requirements.txt
streamlit
crewai
langchain
python-docx
PyPDF2
pandas
openpyxl
python-dotenv
ollama
requests
sqlalchemy
plotly
beautifulsoup4
selenium
webdriver-manager
moviepy
pillow
opencv-python
requests-html
googlesearch-python
geopy
certifi
urllib3